requirements.txt
numpy
pillow
scikit-learn
flask
Python-IO
awsebcli